+++ gpdb-20160503/core/Client.cs
@@ -34,20 +34,20 @@ namespace AlibabaCloud.SDK.Gpdb20160503
                 {"cn-shenzhen-finance-1", "gpdb.aliyuncs.com"},
                 {"cn-qingdao", "gpdb.aliyuncs.com"},
                 {"cn-north-2-gov-1", "gpdb.aliyuncs.com"},
-                {"me-east-1", "gpdb.me-east-1.aliyuncs.com"},
-                {"me-central-1", "gpdb.me-central-1.aliyuncs.com"},
-                {"eu-west-1", "gpdb.eu-west-1.aliyuncs.com"},
-                {"eu-central-1", "gpdb.eu-central-1.aliyuncs.com"},
-                {"cn-zhangjiakou", "gpdb.cn-zhangjiakou.aliyuncs.com"},
                 {"cn-wulanchabu", "gpdb.cn-wulanchabu.aliyuncs.com"},
-                {"cn-huhehaote", "gpdb.cn-huhehaote.aliyuncs.com"},
-                {"cn-chengdu", "gpdb.cn-chengdu.aliyuncs.com"},
-                {"cn-beijing-finance-1", "gpdb.cn-beijing-finance-1.aliyuncs.com"},
-                {"ap-southeast-7", "gpdb.ap-southeast-7.aliyuncs.com"},
-                {"ap-southeast-5", "gpdb.ap-southeast-5.aliyuncs.com"},
-                {"ap-southeast-3", "gpdb.ap-southeast-3.aliyuncs.com"},
+                {"cn-zhangjiakou", "gpdb.cn-zhangjiakou.aliyuncs.com"},
                 {"ap-northeast-2", "gpdb.ap-northeast-2.aliyuncs.com"},
                 {"ap-northeast-1", "gpdb.ap-northeast-1.aliyuncs.com"},
+                {"cn-chengdu", "gpdb.cn-chengdu.aliyuncs.com"},
+                {"ap-southeast-3", "gpdb.ap-southeast-3.aliyuncs.com"},
+                {"cn-huhehaote", "gpdb.cn-huhehaote.aliyuncs.com"},
+                {"ap-southeast-5", "gpdb.ap-southeast-5.aliyuncs.com"},
+                {"ap-southeast-7", "gpdb.ap-southeast-7.aliyuncs.com"},
+                {"eu-west-1", "gpdb.eu-west-1.aliyuncs.com"},
+                {"eu-central-1", "gpdb.eu-central-1.aliyuncs.com"},
+                {"me-east-1", "gpdb.me-east-1.aliyuncs.com"},
+                {"me-central-1", "gpdb.me-central-1.aliyuncs.com"},
+                {"cn-beijing-finance-1", "gpdb.cn-beijing-finance-1.aliyuncs.com"},
             };
             CheckConfig(config);
             this._endpoint = GetEndpoint("gpdb", _regionId, _endpointRule, _network, _suffix, _endpointMap, _endpoint);
@@ -31789,7 +31789,7 @@ namespace AlibabaCloud.SDK.Gpdb20160503
         /// 
         /// <term><b>Description:</b></term>
         /// <description>
-        /// <para>Performs a paged query for all branches under a specified Supabase project. You can filter branches by parent branch, keyword, and sorting criteria. Paging is supported.</para>
+        /// <para>This operation performs a paging query of all branches under a specified Supabase project. You can filter results by parent branch, keyword, and sorting conditions.</para>
         /// </description>
         /// 
         /// <param name="request">
@@ -31868,7 +31868,7 @@ namespace AlibabaCloud.SDK.Gpdb20160503
         /// 
         /// <term><b>Description:</b></term>
         /// <description>
-        /// <para>Performs a paged query for all branches under a specified Supabase project. You can filter branches by parent branch, keyword, and sorting criteria. Paging is supported.</para>
+        /// <para>This operation performs a paging query of all branches under a specified Supabase project. You can filter results by parent branch, keyword, and sorting conditions.</para>
         /// </description>
         /// 
         /// <param name="request">
@@ -31947,7 +31947,7 @@ namespace AlibabaCloud.SDK.Gpdb20160503
         /// 
         /// <term><b>Description:</b></term>
         /// <description>
-        /// <para>Performs a paged query for all branches under a specified Supabase project. You can filter branches by parent branch, keyword, and sorting criteria. Paging is supported.</para>
+        /// <para>This operation performs a paging query of all branches under a specified Supabase project. You can filter results by parent branch, keyword, and sorting conditions.</para>
         /// </description>
         /// 
         /// <param name="request">
@@ -31970,7 +31970,7 @@ namespace AlibabaCloud.SDK.Gpdb20160503
         /// 
         /// <term><b>Description:</b></term>
         /// <description>
-        /// <para>Performs a paged query for all branches under a specified Supabase project. You can filter branches by parent branch, keyword, and sorting criteria. Paging is supported.</para>
+        /// <para>This operation performs a paging query of all branches under a specified Supabase project. You can filter results by parent branch, keyword, and sorting conditions.</para>
         /// </description>
         /// 
         /// <param name="request">