+++ gpdb-20160503/core/Models/ListBranchesResponseBody.cs
@@ -73,6 +73,10 @@ namespace AlibabaCloud.SDK.Gpdb20160503.Models
                 [Validation(Required=false)]
                 public string ServiceType { get; set; }
 
+                [NameInMap("SpbProjectId")]
+                [Validation(Required=false)]
+                public string SpbProjectId { get; set; }
+
                 [NameInMap("Status")]
                 [Validation(Required=false)]
                 public string Status { get; set; }
@@ -102,7 +106,7 @@ namespace AlibabaCloud.SDK.Gpdb20160503.Models
         }
 
         /// <summary>
-        /// <para>The maximum number of records to return in this request.</para>
+        /// <para>The maximum number of records to return in this query.</para>
         /// 
         /// <b>Example:</b>
         /// <para>20</para>
@@ -112,7 +116,7 @@ namespace AlibabaCloud.SDK.Gpdb20160503.Models
         public int? MaxResults { get; set; }
 
         /// <summary>
-        /// <para>The cursor for the paged query. You do not need to specify this parameter for the first request. For subsequent requests, use the NextToken value returned in the previous response for paging.</para>
+        /// <para>The cursor for paging query. You do not need to specify this parameter for the first query. For subsequent queries, use the NextToken value returned in the previous response.</para>
         /// 
         /// <b>Example:</b>
         /// <para>caeba0bbb2be03f84eb48b699f0a4883</para>
@@ -160,7 +164,7 @@ namespace AlibabaCloud.SDK.Gpdb20160503.Models
         public string RequestId { get; set; }
 
         /// <summary>
-        /// <para>The total number of branches that match the query criteria.</para>
+        /// <para>The total number of branches that match the query conditions.</para>
         /// 
         /// <b>Example:</b>
         /// <para>20</para>