+++ gpdb-20160503/ChangeLog.md
@@ -1,3 +1,8 @@
+2026-08-10 Version: 3.12.1
+- Update API DescribeBranch: add response parameters Body.Branch.SpbProjectId.
+- Update API ListBranches: add response parameters Body.Branches.$.SpbProjectId.
+
+
 2026-08-04 Version: 3.12.0
 - Support API CheckSaasServiceVersion.
 - Support API GetApiEndpoints.